merge with head

+++ californium-core/src/test/java/org/eclipse/californium/core/test/lockstep/DeduplicationTest.java
@@ -24,14 +24,12 @@ import static org.eclipse.californium.core.coap.CoAP.ResponseCode.CONTENT;
 import static org.eclipse.californium.core.coap.CoAP.Type.ACK;
 import static org.eclipse.californium.core.coap.CoAP.Type.CON;
 import static org.eclipse.californium.core.coap.CoAP.Type.RST;
+import static org.eclipse.californium.core.test.lockstep.IntegrationTestTools.*;
+import static org.junit.Assert.*;
 
-import java.io.IOException;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 
-import org.junit.Assert;
-
-import org.eclipse.californium.core.coap.CoAP.Code;
 import org.eclipse.californium.category.Medium;
 import org.eclipse.californium.core.coap.Request;
 import org.eclipse.californium.core.coap.Response;
@@ -50,60 +48,50 @@ import org.junit.experimental.categories.Category;
  */
 @Category(Medium.class)
 public class DeduplicationTest {
-	
+
 	private LockstepEndpoint server;
-	
+
 	private Endpoint client;
 	private int clientPort;
-	
+
 	@Before
-	public void setupServer() throws IOException {
-		System.out.println("\nStart "+getClass().getSimpleName());
-		
+	public void setupServer() throws Exception {
+		System.out.println(System.lineSeparator() + "Start " + getClass().getSimpleName());
+
 		NetworkConfig config = new NetworkConfig()
 			.setInt(NetworkConfig.Keys.MAX_MESSAGE_SIZE, 128)
 			.setInt(NetworkConfig.Keys.PREFERRED_BLOCK_SIZE, 128)
 			.setInt(NetworkConfig.Keys.ACK_TIMEOUT, 200) // client retransmits after 200 ms
 			.setInt(NetworkConfig.Keys.ACK_RANDOM_FACTOR, 1);
-		client = new CoapEndpoint(new InetSocketAddress(0), config);
+		client = new CoapEndpoint(new InetSocketAddress(InetAddress.getLoopbackAddress(), 0), config);
 		client.addInterceptor(new MessageTracer());
 		client.start();
 		clientPort = client.getAddress().getPort();
-		System.out.println("Client binds to port "+clientPort);
+		server = createLockstepEndpoint(client.getAddress());
+		System.out.println("Client binds to port " + clientPort);
 	}
-	
+
 	@After
 	public void shutdownServer() {
-		System.out.println();
-		client.destroy();
-		System.out.println("End "+getClass().getSimpleName());
-	}
-	
-	@Test
-	public void test() throws Throwable {
-		try {
-			testGET();
-			
-		} catch (Exception e) {
-			e.printStackTrace();
-			System.err.println("Make sure you did not forget a .go() at the end of a line.");
-			throw e;
-		} catch (Throwable t) {
-			System.err.println(t);
-			throw t;
+		if (server != null) {
+			server.destroy();
 		}
+		if (client != null) {
+			client.destroy();
+		}
+		System.out.println(System.lineSeparator() + "End " + getClass().getSimpleName());
 	}
-	
-	private void testGET() throws Exception {
-		System.out.println("Simple blockwise GET:");
+
+	@Test
+	public void testGET() throws Exception {
+		System.out.println("Simple GET:");
 		String path = "test";
 		String payload = "possible conflict";
-		server = createLockstepEndpoint();
-		
-		Request request = createRequest(GET, path);
+
+		Request request = createRequest(GET, path, server);
 		request.setMID(1234);
 		client.sendRequest(request);
-		
+
 		server.expectRequest(CON, GET, path).storeToken("A").go();
 		server.sendEmpty(ACK).mid(1234).go();
 		server.sendEmpty(ACK).mid(1234).go();
@@ -113,38 +101,19 @@ public class DeduplicationTest {
 		server.expectEmpty(ACK, 4711).go();
 		server.sendResponse(CON, CONTENT).loadToken("A").mid(42).payload("separate").go();
 		server.expectEmpty(RST, 42).go();
-		
-		request = createRequest(GET, path);
+
+		request = createRequest(GET, path, server);
 		request.setMID(4711);
 		client.sendRequest(request);
-		
+
 		server.expectRequest(CON, GET, path).storeBoth("B").storeToken("C").go();
 		server.sendResponse(ACK, CONTENT).loadBoth("B").payload("possible conflict").go();
 		server.sendResponse(ACK, CONTENT).loadBoth("B").payload("possible conflict").go();
-		
-		Response response = request.waitForResponse(1000);
-		Assert.assertNotNull("Client received no response", response);
-		Assert.assertEquals("Client received wrong response code:", CONTENT, response.getCode());
-		Assert.assertEquals("Client received wrong payload:", payload, response.getPayloadString());
-		
-		response = request.waitForResponse(1000);
-		Assert.assertNull("Client received duplicate", response);
-	}
-		
-	private LockstepEndpoint createLockstepEndpoint() {
-		try {
-			LockstepEndpoint endpoint = new LockstepEndpoint();
-			endpoint.setDestination(new InetSocketAddress(InetAddress.getByName("localhost"), clientPort));
-			return endpoint;
-		} catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-	
-	private Request createRequest(Code code, String path) throws Exception {
-		Request request = new Request(code);
-		String uri = "coap://localhost:"+(server.getPort())+"/"+path;
-		request.setURI(uri);
-		return request; 
+
+		Response response = request.waitForResponse(500);
+		assertResponseContainsExpectedPayload(response, CONTENT, payload);
+
+		response = request.waitForResponse(500);
+		assertNull("Client received duplicate", response);
 	}
 }
